added data if exists for userId and userName

+++ src/main/java/org/workshop/microphoneschedulerapi/domain/dto/CustomMicrophoneListDTO.java
@@ -4,5 +4,5 @@ import lombok.Builder;
 
 @Builder
 public record CustomMicrophoneListDTO(int sceneId, String sceneName, Integer personageId, String personageName,
-                                      int userId, String userName, Integer microphoneId, String microphoneName) {
+                                      Long userId, String userName, Integer microphoneId, String microphoneName) {
 }
